Polish function code
