+++ paddle/fluid/jit/all.h
@@ -14,7 +14,7 @@
 
 #pragma once
 
-#include "function/base_function.h"  // NOLINT
-#include "layer.h"                   // NOLINT
-#include "serializer.h"              // NOLINT
-#include "serializer_utils.h"        // NOLINT
+#include "engine/base_engine.h"  // NOLINT
+#include "layer.h"               // NOLINT
+#include "serializer.h"          // NOLINT
+#include "serializer_utils.h"    // NOLINT
