+++ paddle/fluid/jit/compilation_unit.cc
@@ -16,7 +16,7 @@
 
 #include "paddle/phi/core/enforce.h"
 
-#include "paddle/fluid/jit/function/base_function.h"
+#include "paddle/fluid/jit/engine/base_engine.h"
 
 namespace paddle {
 namespace jit {
