+++ paddle/fluid/jit/engine/CMakeLists.txt
@@ -1,0 +1,9 @@
+cc_library(
+  jit_executor_engine
+  SRCS executor_engine.cc
+  DEPS executor)
+
+cc_library(
+  jit_pe_engine
+  SRCS pe_engine.cc
+  DEPS parallel_executor)
