+++ paddle/fluid/jit/engine/executor_engine.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "paddle/fluid/jit/function/executor_function.h"
+#include "paddle/fluid/jit/engine/executor_engine.h"
 
 #include "paddle/fluid/framework/program_desc.h"
 #include "paddle/fluid/framework/variable.h"
