+++ paddle/fluid/jit/engine/pe_engine.h
@@ -16,23 +16,37 @@
 
 #include <vector>
 
-#include "paddle/fluid/framework/executor.h"
 #include "paddle/fluid/framework/scope.h"
 
-#include "paddle/fluid/jit/function/base_function.h"
+#include "paddle/fluid/jit/engine/base_engine.h"
 #include "paddle/fluid/jit/function_schema.h"
 #include "paddle/fluid/jit/function_utils.h"
 
 namespace paddle {
+namespace framework {
+class ParallelExecutor;
+namespace details {
+class ExecutionStrategy;
+}
+namespace ir {
+class Graph;
+}
+}  // namespace framework
+
 namespace jit {
+using ExecutionStrategy = framework::details::ExecutionStrategy;
+using ParallelExecutor = framework::ParallelExecutor;
+using Graph = framework::ir::Graph;
 
-class ExecutorFunction : public BaseFunction {
+class PEFunction : public BaseFunction {
  public:
-  ExecutorFunction(const std::shared_ptr<FunctionInfo> &info,
-                   const Name2VariableMap &params_dict,
-                   const phi::Place &place);
+  PEFunction(const std::shared_ptr<FunctionInfo> &info,
+             const Name2VariableMap &params_dict,
+             const phi::Place &place);
 
-  ~ExecutorFunction() noexcept {}
+  ~PEFunction() noexcept {}
+
+  void CreateGraphAndPE();
 
   std::vector<Tensor> operator()(const std::vector<Tensor> &inputs);
 
@@ -44,7 +58,8 @@ class ExecutorFunction : public BaseFunction {
   std::shared_ptr<FunctionInfo> info_;
   framework::Scope scope_;
   phi::Place place_;
-  framework::Executor inner_exe_;
+  std::shared_ptr<ParallelExecutor> inner_pe_;
+  std::shared_ptr<Graph> graph_;
 };
 
 }  // namespace jit
