+++ paddle/fluid/jit/layer.h
@@ -21,7 +21,7 @@
 #include "paddle/phi/api/include/tensor.h"
 #include "paddle/phi/common/place.h"
 
-#include "function/base_function.h"  //NOLINT
+#include "engine/base_engine.h"  //NOLINT
 
 namespace paddle {
