+++ paddle/fluid/pybind/jit.cc
@@ -18,8 +18,8 @@ limitations under the License. */
 #include "paddle/fluid/imperative/layer.h"
 #include "paddle/fluid/platform/place.h"
 
-#include "paddle/fluid/jit/function/executor_function.h"
-#include "paddle/fluid/jit/function/pe_function.h"
+#include "paddle/fluid/jit/engine/executor_engine.h"
+#include "paddle/fluid/jit/engine/pe_engine.h"
 #include "paddle/fluid/jit/function_schema.h"
 #include "paddle/fluid/jit/layer.h"
 #include "paddle/fluid/jit/serializer.h"
@@ -29,8 +29,8 @@ namespace py = pybind11;
 namespace paddle {
 namespace pybind {
 
-PyTypeObject *g_executor_function_pytype = nullptr;
-PyTypeObject *g_pe_function_pytype = nullptr;
+PyTypeObject *g_executor_engine_pytype = nullptr;
+PyTypeObject *g_pe_engine_pytype = nullptr;
 using Variable = paddle::framework::Variable;
 
 void BindJit(pybind11::module *m) {
@@ -40,16 +40,15 @@ void BindJit(pybind11::module *m) {
            py::return_value_policy::reference);
 
   py::class_<jit::ExecutorFunction, std::shared_ptr<jit::ExecutorFunction>>
-      executor_function(
-          *m, "ExectorFunction", R"DOC(ExectorFunction Class.)DOC");
-  g_executor_function_pytype =
-      reinterpret_cast<PyTypeObject *>(executor_function.ptr());
-  executor_function.def("info", &jit::ExecutorFunction::Info);
+      executor_engine(*m, "ExectorFunction", R"DOC(ExectorFunction Class.)DOC");
+  g_executor_engine_pytype =
+      reinterpret_cast<PyTypeObject *>(executor_engine.ptr());
+  executor_engine.def("info", &jit::ExecutorFunction::Info);
 
-  py::class_<jit::PEFunction, std::shared_ptr<jit::PEFunction>> pe_function(
+  py::class_<jit::PEFunction, std::shared_ptr<jit::PEFunction>> pe_engine(
       *m, "PEFunction", R"DOC(PEFunction Class.)DOC");
-  g_pe_function_pytype = reinterpret_cast<PyTypeObject *>(pe_function.ptr());
-  pe_function.def("info", &jit::PEFunction::Info);
+  g_pe_engine_pytype = reinterpret_cast<PyTypeObject *>(pe_engine.ptr());
+  pe_engine.def("info", &jit::PEFunction::Info);
 
   py::class_<jit::FunctionInfo, std::shared_ptr<jit::FunctionInfo>>(
       *m, "FunctionInfo", R"DOC(FunctionInfo Class.)DOC")
